Create a task.install.util submodule
This will hold miscellaneous code that doesn't need to be top-level and
isn't distro-specific
